Proyecto Final

+++ L07 Deliverable - Deployment/Acme-HandyWork/src/main/java/repositories/CustomisationSystemRepository.java
@@ -42,16 +42,16 @@ public interface CustomisationSystemRepository extends JpaRepository<Customisati
 
 	//Querys --> Ratio of applications 12.5.5
 	@Query("select (select count(a) from Application a where a.status = 'pending')*1.0/count(ap) from Application ap")
-	double findRatioOfPendingApplications();
+	Long findRatioOfPendingApplications();
 	//Querys --> Ratio of applications 12.5.6
 	@Query("select count(f.status)*1.0/(select count(t)*1.0 from Application t) from Application f where f.status = 'accepted'")
-	double findRationOfAcceptedAplications();
+	Long findRationOfAcceptedAplications();
 	//Querys --> Ratio of applications 12.5.7
 	@Query("select count(f.status)*1.0/(select count(t)*1.0 from Application t) from Application f where f.status = 'rejected'")
-	double findRationOfRejectedApplications();
+	Long findRationOfRejectedApplications();
 	//Querys --> Ratio of applications 12.5.8
 	@Query("select (count(a)*1.0/(select count(ap) from Application ap)) from Application a where current_date() > a.momentElapsed and a.status = 'pending'")
-	double findRationOfPendingApplicationCannotChangeItsStatus();
+	Long findRationOfPendingApplicationCannotChangeItsStatus();
 
 	//Querys --> Complaints Per FixUpTasks --> NO DASHBOARD
 	@Query("select avg(f.complaint.size), min(f.complaint.size), max(f.complaint.size), stddev(f.complaint.size) from FixUpTask f")
@@ -63,7 +63,7 @@ public interface CustomisationSystemRepository extends JpaRepository<Customisati
 
 	//The ratio of fix-up-tasks with a complaint
 	@Query("select count(a)*1.0/(select count(t)*1.0 from FixUpTask t) from FixUpTask a where a.complaint.size > 0")
-	double ratioOfFixUpTasksWithComplaint();
+	Long ratioOfFixUpTasksWithComplaint();
 
 	//Querys --> 10% above the average 12.5.9
 	@Query("select f from Customer f join f.fixUpTask t where f.fixUpTask.size > (select avg(f.fixUpTask.size)+(avg(f.fixUpTask.size)/10)*1.0 from Customer f) order by t.application.size")
